nullability: fix edge case when `NULL` is used
Since `NULL` is a macro, when transitioning nullability semantics, the
usual definition of `((void *)0)` has different nullability qualifiers
attached to it, depending on context. Thus, when under NNBD-strict
semantics, we get a cast error.

We fix this by circumventing the nullability checks in `castexpr()` when
the expression is a null pointer constant.