makefile: fix completely broken prerequisite on build directory
A directory's timestamp is updated when a new file is added. Thus, the
build directory was 'newer' than some of object files it contained,
leading `make` to rebuild them, since their creation depends upon the
existence of the build directory, otherwise we wouldn't be able to store
them there. Traversing the dependency graph, everything was rebuild
everytime, even without any modifications.

We fix this by creating a dummy `marker` file inside the build
directory. The object files now depend on this file, instead of the
directory. This works because the directory is created alongside the
`marker` (one `marker` file for each build directory) and files have
stable timestamps, as long as they're not modified.